refactor: combine diff tests
Move diff tests from test_009diff.py into test008diff.py so that all tests for Diff and diff_models are in one location
