+++ drivers/python/tests/test_008diff.py
@@ -1,13 +1,15 @@
+"""
+Tests for diff.py module
+"""
 import os
 import sys
 
+from bento_mdf.diff import Diff, diff_models
+from bento_mdf.mdf import MDF
+from bento_meta.objects import Property, Term, ValueSet
+
 sys.path.insert(0, ".")
 sys.path.insert(0, "..")
-from bento_mdf.diff import diff_models
-
-# from pdb import set_trace
-from bento_mdf.mdf import MDF
-
 tdir = "tests/" if os.path.exists("tests") else ""
 
 
@@ -25,7 +27,6 @@ def test_diff_of_extra_node_properties_and_terms():
     a = MDF(tdir + "samples/test-model-a.yml", handle="test")
     b = MDF(tdir + "samples/test-model-b.yml", handle="test")
     actual = diff_models(a.model, b.model)
-    # expected = {'nodes': {'file': {'props': {'a': set(), 'b': {'encryption_type'}}}}, 'props': {('sample', 'sample_type'): {'value_set': {'a': set(), 'b': {'not a tumor'}}}, 'a': set(), 'b': {('file', 'encryption_type')}}}
     expected = {
         "nodes": {"file": {"props": {"a": None, "b": ["encryption_type"]}}},
         "props": {
@@ -42,7 +43,6 @@ def test_diff_of_extra_node_property():
     a = MDF(tdir + "samples/test-model-a.yml", handle="test")
     b = MDF(tdir + "samples/test-model-d.yml", handle="test")
     actual = diff_models(a.model, b.model)
-    # expected = {'nodes': {'diagnosis': {'props': {'a': set(), 'b': {'fatal'}}}}, 'props': {'a': set(), 'b': {('diagnosis', 'fatal')}}}
     expected = {
         "nodes": {"diagnosis": {"props": {"a": None, "b": ["fatal"]}}},
         "props": {"a": None, "b": [("diagnosis", "fatal")]},
@@ -55,7 +55,6 @@ def test_diff_of_extra_node_edge_and_property():
     a = MDF(tdir + "samples/test-model-a.yml", handle="test")
     b = MDF(tdir + "samples/test-model-e.yml", handle="test")
     actual = diff_models(a.model, b.model)
-    # expected = {'nodes': {'a': set(), 'b': {'outcome'}}, 'edges': {'a': set(), 'b': {('end_result', 'diagnosis', 'outcome')}}, 'props': {'a': set(), 'b': {('outcome', 'fatal')}}}
     expected = {
         "nodes": {"a": None, "b": ["outcome"]},
         "edges": {"a": None, "b": [("end_result", "diagnosis", "outcome")]},
@@ -69,7 +68,6 @@ def test_diff_of_extra_node():
     a = MDF(tdir + "samples/test-model-a.yml", handle="test")
     b = MDF(tdir + "samples/test-model-f.yml", handle="test")
     actual = diff_models(a.model, b.model)
-    # expected = {'nodes': {'a': {'diagnosis'}, 'b': set()}, 'edges': {'a': {('of_case', 'diagnosis', 'case')}, 'b': set()}, 'props': {'a': {('diagnosis', 'disease')}, 'b': set()}}
     expected = {
         "nodes": {"a": ["diagnosis"], "b": None},
         "edges": {"a": [("of_case", "diagnosis", "case")], "b": None},
@@ -100,8 +98,6 @@ def test_diff_of_swapped_nodeprops():
         node_val["props"]["b"].sort()
     actual["props"]["a"].sort()
     actual["props"]["b"].sort()
-    # expected = {'nodes': {'file': {'props': {'a': {'file_name', 'file_size', 'md5sum'}, 'b': {'disease'}}}, 'diagnosis': {'props': {'a': {'disease'}, 'b': {'file_name', 'file_size', 'md5sum'}}}}, 'props': {'a': {('file', 'file_name'), ('file', 'file_size'), ('diagnosis', 'disease'), ('file', 'md5sum')}, 'b': {('diagnosis', 'file_name'), ('file', 'disease'), ('diagnosis', 'file_size'), ('diagnosis', 'md5sum')}}}
-    # expected = {'nodes': {'file': {'props': {'a': ['file_name', 'file_size', 'md5sum'], 'b': ['disease']}}, 'diagnosis': {'props': {'a': ['disease'], 'b': ['file_name', 'file_size', 'md5sum']}}}, 'props': {'a': [('file', 'file_name'), ('file', 'file_size'), ('diagnosis', 'disease'), ('file', 'md5sum')], 'b': [('diagnosis', 'file_name'), ('file', 'disease'), ('diagnosis', 'file_size'), ('diagnosis', 'md5sum')]}}
     expected = {
         "nodes": {
             "diagnosis": {
@@ -145,7 +141,6 @@ def test_diff_of_assorted_changes():
     a = MDF(tdir + "samples/test-model-d.yml", handle="test")
     b = MDF(tdir + "samples/test-model-e.yml", handle="test")
     actual = diff_models(a.model, b.model)
-    # expected = {'nodes': {'diagnosis': {'props': {'a': {'fatal'}, 'b': set()}}, 'a': set(), 'b': {'outcome'}}, 'edges': {'a': set(), 'b': {('end_result', 'diagnosis', 'outcome')}}, 'props': {'a': {('diagnosis', 'fatal')}, 'b': {('outcome', 'fatal')}}}
     expected = {
         "nodes": {
             "diagnosis": {"props": {"a": ["fatal"], "b": None}},
@@ -155,4 +150,134 @@ def test_diff_of_assorted_changes():
         "edges": {"a": None, "b": [("end_result", "diagnosis", "outcome")]},
         "props": {"a": [("diagnosis", "fatal")], "b": [("outcome", "fatal")]},
     }
+    assert actual == expected
+
+
+diff = Diff()
+term_a = Term({"value": "Merida"})
+term_b = Term({"value": "Cumana"})
+term_c = Term({"value": "Maracaibo"})
+term_d = Term({"value": "Ciudad Bolivar"})
+term_e = Term({"value": "Barcelona"})
+term_f = Term({"value": "Barquisimeto"})
+
+
+def test_valuesets_are_different__a():
+    """test using sets as input"""
+
+    vs_1 = ValueSet({"_id": "1"})
+    vs_2 = ValueSet({"_id": "2"})
+
+    vs_1.terms["Merida"] = term_a
+    vs_1.terms["Cumana"] = term_b
+    vs_1.terms["Maracaibo"] = term_c
+    vs_1.terms["Ciudad Bolivar"] = term_d
+
+    vs_2.terms["Merida"] = term_a
+    vs_2.terms["Cumana"] = term_b
+    vs_2.terms["Maracaibo"] = term_c
+    vs_2.terms["Ciudad Bolivar"] = term_d
+
+    actual = diff.valuesets_are_different(vs_1, vs_2)
+    expected = False
+    assert actual == expected
+
+
+def test_valuesets_are_different__b():
+    """test using sets as input"""
+
+    vs_1 = ValueSet({"_id": "1"})
+    vs_2 = ValueSet({"_id": "2"})
+
+    vs_1.terms["Merida"] = term_a
+    vs_1.terms["Cumana"] = term_b
+    vs_1.terms["Maracaibo"] = term_c
+    vs_1.terms["Ciudad Bolivar"] = term_d
+
+    vs_2.terms["Merida"] = term_a
+    vs_2.terms["Cumana"] = term_b
+    vs_2.terms["Maracaibo"] = term_c
+
+    actual = diff.valuesets_are_different(vs_1, vs_2)
+    expected = True
+    assert actual == expected
+
+
+def test_valuesets_are_different__c():
+    """test using sets as input"""
+
+    vs_1 = ValueSet({"_id": "1"})
+    vs_2 = ValueSet({"_id": "2"})
+
+    vs_1.terms["Merida"] = term_a
+    vs_1.terms["Cumana"] = term_b
+    vs_1.terms["Maracaibo"] = term_c
+    vs_1.terms["Ciudad Bolivar"] = term_d
+
+    vs_2.terms["Merida"] = term_a
+    vs_2.terms["Cumana"] = term_b
+    vs_2.terms["Maracaibo"] = term_c
+    vs_2.terms["Ciudad Bolivar"] = term_d
+    vs_2.terms["Barcelona"] = term_e
+
+    actual = diff.valuesets_are_different(vs_1, vs_2)
+    expected = True
+    assert actual == expected
+
+
+def test_valuesets_are_different__d():
+    """test using sets as input"""
+
+    vs_1 = ValueSet({"_id": "1"})
+    vs_2 = ValueSet({"_id": "2"})
+
+    actual = diff.valuesets_are_different(vs_1, vs_2)
+    expected = False
+    assert actual == expected
+
+
+def test_valuesets_are_different__e():
+    """test using sets as input"""
+
+    vs_1 = ValueSet({"_id": "1"})
+
+    vs_1.terms["Merida"] = term_a
+    vs_1.terms["Cumana"] = term_b
+    vs_1.terms["Maracaibo"] = term_c
+    vs_1.terms["Ciudad Bolivar"] = term_d
+
+    actual = diff.valuesets_are_different(vs_1, vs_1)
+    expected = False
+    assert actual == expected
+
+
+def test_valuesets_are_different__f():
+    """test using sets as input"""
+    p_1 = Property({"handle": "States"})
+    p_2 = Property({"handle": "Estados"})
+    vs_1 = ValueSet({"_id": "1"})
+    vs_2 = ValueSet({"_id": "2"})
+
+    term_a = Term({"value": "Merida"})
+    term_b = Term({"value": "Cumana"})
+    term_c = Term({"value": "Maracaibo"})
+    term_d = Term({"value": "Ciudad Bolivar"})
+    term_e = Term({"value": "Barcelona"})
+    term_f = Term({"value": "Barquisimeto"})
+
+    vs_1.terms["Merida"] = term_a
+    vs_1.terms["Cumana"] = term_b
+    vs_1.terms["Maracaibo"] = term_c
+    vs_1.terms["Ciudad Bolivar"] = term_d
+
+    vs_2.terms["Merida"] = term_a
+    vs_2.terms["Cumana"] = term_b
+    vs_2.terms["Maracaibo"] = term_c
+    vs_2.terms["Ciudad Bolivar"] = term_d
+
+    p_1.value_set = vs_1
+    p_2.value_set = vs_2
+
+    actual = diff.valuesets_are_different(p_1.value_set, p_2.value_set)
+    expected = False
     assert actual == expected
